# tests/fixtures/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Interactive Note Questions API
  version: 1.0.0
  description: >-
    Authoritative local API for workspaces, notes, shared questions, organization,
    search, reminders, safe deletion, and portable import/export. All timestamps
    are RFC 3339 UTC unless a schema explicitly uses a calendar date.
servers:
  - url: /
    description: Same-origin local application
security: []
...
paths:
...
  /api/v1/workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      responses:
        '200':
          description: All workspaces
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Workspace' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/workspaces/{workspaceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Workspaces]
      operationId: getWorkspace
      responses:
        '200':
          description: Workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Error' }
    put:
      tags: [Workspaces]
      operationId: updateWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, version]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                version: { $ref: '#/components/schemas/Version' }
      responses:
        '200':
          description: Updated workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/topics:
    get:
      tags: [Topics]
      operationId: listTopics
      parameters:
        - $ref: '#/components/parameters/RequiredWorkspaceQuery'
      responses:
        '200':
          description: Topics in workspace
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Topic' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Topics]
      operationId: createTopic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [workspaceId, name]
              properties:
                workspaceId: { $ref: '#/components/schemas/Id' }
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '201':
          description: Topic created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Topic' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/topics/{topicId}:
    parameters:
      - name: topicId
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    put:
      tags: [Topics]
      operationId: updateTopic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, version]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                version: { $ref: '#/components/schemas/Version' }
      responses:
        '200':
          description: Updated topic
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Topic' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      tags: [Topics]
      operationId: deleteTopic
      parameters:
        - $ref: '#/components/parameters/IfVersion'
      responses:
        '204': { description: Topic deleted and note assignments cleared }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/tags:
    get:
      tags: [Tags]
      operationId: listTags
      parameters:
        - $ref: '#/components/parameters/RequiredWorkspaceQuery'
        - name: includeShared
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: Owned and available tags
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Tag' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Tags]
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ownerWorkspaceId, name]
              properties:
                ownerWorkspaceId: { $ref: '#/components/schemas/Id' }
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/tags/{tagId}:
    parameters:
      - name: tagId
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    put:
      tags: [Tags]
      operationId: updateTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name, version]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                version: { $ref: '#/components/schemas/Version' }
      responses:
        '200':
          description: Updated tag
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/tags/{tagId}/workspace-access:
    parameters:
      - name: tagId
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Id' }
    put:
      tags: [Tags]
      operationId: setTagWorkspaceAccess
      description: Replaces the explicit non-owner workspace access set.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [workspaceIds, version, removalDecisions]
              properties:
                workspaceIds:
                  type: array
                  uniqueItems: true
                  items: { $ref: '#/components/schemas/Id' }
                version: { $ref: '#/components/schemas/Version' }
                removalDecisions:
                  type: array
                  description: Required for assignments made inaccessible by this update.
                  items:
                    type: object
                    additionalProperties: false
                    required: [workspaceId, removeAssignments]
                    properties:
                      workspaceId: { $ref: '#/components/schemas/Id' }
                      removeAssignments: { type: boolean }
      responses:
        '200':
          description: Updated tag with access list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/notes:
    get:
      tags: [Notes]
      operationId: listNotes
      parameters:
        - $ref: '#/components/parameters/RequiredWorkspaceQuery'
        - name: parentNoteId
          in: query
          description: Omit for all notes; use literal `root` for root notes.
          schema: { type: string }
        - name: topicId
          in: query
          schema: { $ref: '#/components/schemas/Id' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Notes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotePage' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Notes]
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NoteWrite' }
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/notes/{noteId}:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    get:
      tags: [Notes]
      operationId: getNote
      responses:
        '200':
          description: Note with linked question summaries
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Error' }
    put:
      tags: [Notes]
      operationId: updateNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/NoteWrite'
                - type: object
                  required: [version]
                  properties:
                    version: { $ref: '#/components/schemas/Version' }
      responses:
        '200':
          description: Updated note and links
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Note' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/notes/{noteId}/deletion-preview:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    post:
      tags: [Notes]
      operationId: previewNoteDeletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [version]
              properties:
                version: { $ref: '#/components/schemas/Version' }
      responses:
        '200':
          description: Current impact and short-lived preview token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeletionPreview' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/notes/{noteId}/delete:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    post:
      tags: [Notes]
      operationId: executeNoteDeletion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeleteNoteCommand' }
      responses:
        '204': { description: Reviewed deletion committed atomically }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/questions:
    get:
      tags: [Questions]
      operationId: listQuestions
      parameters:
        - $ref: '#/components/parameters/RequiredWorkspaceQuery'
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/QuestionStatus' }
          explode: false
        - name: kind
          in: query
          schema: { $ref: '#/components/schemas/QuestionKind' }
        - name: topicId
          in: query
          schema: { $ref: '#/components/schemas/Id' }
        - name: tagId
          in: query
          schema: { $ref: '#/components/schemas/Id' }
        - name: createdFrom
          in: query
          schema: { type: string, format: date-time }
        - name: createdTo
          in: query
          schema: { type: string, format: date-time }
        - name: updatedFrom
          in: query
          schema: { type: string, format: date-time }
        - name: updatedTo
          in: query
          schema: { type: string, format: date-time }
        - name: dueFrom
          in: query
          schema: { type: string, format: date }
        - name: dueTo
          in: query
          schema: { type: string, format: date }
        - name: hasAnswer
          in: query
          schema: { type: boolean }
        - name: hasLinkedNotes
          in: query
          schema: { type: boolean }
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, dueDate, priority, questionText, status]
            default: updatedAt
        - name: direction
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Filtered question page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuestionPage' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Questions]
      operationId: createQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuestionWrite' }
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Question' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/QuestionId'
    get:
      tags: [Questions]
      operationId: getQuestion
      responses:
        '200':
          description: Canonical question with links
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Question' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Error' }
    put:
      tags: [Questions]
      operationId: updateQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/QuestionWrite'
                - type: object
                  required: [version]
                  properties:
                    version: { $ref: '#/components/schemas/Version' }
      responses:
        '200':
          description: Updated canonical question
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Question' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/search:
    get:
      tags: [Search]
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 500 }
        - name: contentScope
          in: query
          required: true
          schema: { type: string, enum: [notes, questions, answers, everything] }
        - name: workspaceScope
          in: query
          required: true
          schema: { type: string, enum: [current, all] }
        - name: workspaceId
          in: query
          description: Required when workspaceScope is current.
          schema: { $ref: '#/components/schemas/Id' }
        - name: tagId
          in: query
          description: Restrict note results to notes assigned this tag. The predicate uses note-tag existence semantics and never duplicates a note.
          schema: { $ref: '#/components/schemas/Id' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Ranked search results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchPage' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/export:
    post:
      tags: [Data portability]
      operationId: exportAllData
      responses:
        '200':
          description: Complete versioned export ZIP
          headers:
            Content-Disposition:
              schema: { type: string }
          content:
            application/zip:
              schema: { type: string, contentEncoding: binary }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/imports/validate:
    post:
      tags: [Data portability]
      operationId: validateImport
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [archive]
              properties:
                archive:
                  type: string
                  contentMediaType: application/zip
                  contentEncoding: binary
      responses:
        '200':
          description: Valid import preview with any conflicts
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportPreview' }
        '413':
          description: Archive exceeds configured limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/imports/{importId}/apply:
    parameters:
      - $ref: '#/components/parameters/ImportId'
    post:
      tags: [Data portability]
      operationId: applyImport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [conflictResolutions]
              properties:
                conflictResolutions:
                  type: array
                  items: { $ref: '#/components/schemas/ConflictResolution' }
      responses:
        '200':
          description: Import committed
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [importId, state, appliedCounts]
                properties:
                  importId: { $ref: '#/components/schemas/Id' }
                  state: { type: string, const: applied }
                  appliedCounts:
                    type: object
                    additionalProperties: { type: integer, minimum: 0 }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Error' }
  /api/v1/imports/{importId}:
    parameters:
      - $ref: '#/components/parameters/ImportId'
    delete:
      tags: [Data portability]
      operationId: cancelImport
      responses:
        '204': { description: Staged import removed or marked cancelled }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Error' }
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
    RequiredWorkspaceQuery:
      name: workspaceId
      in: query
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    IfVersion:
      name: version
      in: query
      required: true
      schema: { $ref: '#/components/schemas/Version' }
...
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Input failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: Version, uniqueness, or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Id:
      type: string
      format: uuid
      pattern: '^[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}$'
    Version:
      type: integer
      minimum: 1
...
    Workspace:
      type: object
      additionalProperties: false
      required: [id, name, createdAt, updatedAt, version]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: string }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
        version: { $ref: '#/components/schemas/Version' }
    Topic:
      type: object
      additionalProperties: false
      required: [id, workspaceId, name, createdAt, updatedAt, version]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        workspaceId: { $ref: '#/components/schemas/Id' }
        name: { type: string }
        createdAt: { $ref: '#/components/schemas/Timestamp' }
        updatedAt: { $ref: '#/components/schemas/Timestamp' }
        version: { $ref: '#/components/schemas/Version' }
...